requirements.txt
streamlit
ollama
groq
watchdog
python-dotenv